+++ src/BDSCrystalFactory.cc
@@ -16,6 +16,7 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 */
+#include "BDSAcceleratorModel.hh"
 #include "BDSColours.hh"
 #include "BDSCrystal.hh"
 #include "BDSCrystalFactory.hh"
@@ -27,6 +28,7 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "globals.hh"
 #include "G4Box.hh"
+#include "G4DisplacedSolid.hh"
 #include "G4ExtrudedSolid.hh"
 #include "G4LogicalVolume.hh"
 #include "G4Material.hh"
@@ -43,6 +45,7 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 
 #include <algorithm>
 #include <cmath>
+#include <set>
 #include <vector>
 
 // only use the crystal extensions if using 10.4.p00 upwards
@@ -142,6 +145,8 @@ void BDSCrystalFactory::CommonConstruction(const G4String&       nameIn,
   crystalLV = new G4LogicalCrystalVolume(crystalSolid,
 					 crystalMat,
 					 nameIn + "_crystal_lv");
+
+  BDSAcceleratorModel::Instance()->VolumeSet("crystals")->insert(crystalLV);
 #else
   // build logical volumes
   crystalLV = new G4LogicalVolume(crystalSolid,
@@ -185,6 +190,7 @@ BDSCrystal* BDSCrystalFactory::BuildCrystalObject(const BDSExtent& extent)
   aCrystal->RegisterSensitiveVolume(crystalLV);
   aCrystal->RegisterUserLimits(allUserLimits);
   aCrystal->RegisterVisAttributes(allVisAttributes);
+  aCrystal->ExcludeLogicalVolumeFromBiasing(crystalLV); // can't double bias one LV ie with generic biasing
   
   return aCrystal;
 }
@@ -263,19 +269,30 @@ BDSCrystal* BDSCrystalFactory::CreateCrystalCylinder(const G4String&       nameI
   G4double sweepAngle;
   CalculateSolidAngles(ba, startAngle, sweepAngle);
   
-  crystalSolid = new G4Tubs(nameIn + "_solid",
-			    xBR - 0.5*thickness,
-			    xBR + 0.5*thickness,
-			    (recipe->lengthZ)*0.5,
-			    startAngle,
-			    sweepAngle);
+  G4Tubs* rawShape = new G4Tubs(nameIn + "_solid",
+				xBR - 0.5*thickness,
+				xBR + 0.5*thickness,
+				recipe->lengthY*0.5,
+				startAngle,
+				sweepAngle);
 
+  // crystal channelling only works in local unitX of a given solid which
+  // makes it impossible to use a cylinder. we cheat by using a G4DisplacedSolid
+  // that's a class used internally by geant4's boolean solids to rotate and translate
+  // the frame of a solid. another option was an intersection with a big box, but
+  // geant4 can't handle this.
+  G4RotationMatrix* relativeRotation = new G4RotationMatrix();
+  relativeRotation->rotateX(-CLHEP::halfpi);
+  G4ThreeVector offset(0,0,0);
+  crystalSolid = new G4DisplacedSolid(nameIn + "_shifted_solid",
+				      rawShape,
+				      relativeRotation,
+				      offset);
+  
   CommonConstruction(nameIn, recipe);
 
-  // placement transform
+  // placement offset - no rotation as we've rotated the solid internally
   placementOffset = G4ThreeVector(-BendingRadiusHorizontal(recipe), 0, 0);
-  placementRotation = new G4RotationMatrix();
-  placementRotation->rotateX(-CLHEP::halfpi);
 
   BDSExtent ext = CalculateExtents(ba, xBR, thickness, recipe);
   
@@ -344,16 +361,27 @@ BDSCrystal* BDSCrystalFactory::CreateCrystalTorus(const G4String&       nameIn,
       zSections.emplace_back(G4ExtrudedSolid::ZSection(z, G4TwoVector(x, 0), 1));
     }
 
-  crystalSolid = new G4ExtrudedSolid(nameIn + "_solid",
-				     points,
-				     zSections);
+  G4ExtrudedSolid* rawShape = new G4ExtrudedSolid(nameIn + "_solid",
+						  points,
+						  zSections);
+  
+  // crystal channelling only works in local unitX of a given solid which
+  // makes it impossible to use a cylinder. we cheat by using a G4DisplacedSolid
+  // that's a class used internally by geant4's boolean solids to rotate and translate
+  // the frame of a solid. another option was an intersection with a big box, but
+  // geant4 can't handl this.
+  G4RotationMatrix* relativeRotation = new G4RotationMatrix();
+  relativeRotation->rotateX(-CLHEP::halfpi);
+  G4ThreeVector offset(0,0,0);
+  crystalSolid = new G4DisplacedSolid(nameIn + "_shifted_solid",
+				      rawShape,
+				      relativeRotation,
+				      offset);
 
   CommonConstruction(nameIn, recipe);
 
-  // placement transform
+  // placement offset - no rotation as we've rotated the solid internally
   placementOffset = G4ThreeVector(-BendingRadiusHorizontal(recipe), 0, 0);
-  placementRotation = new G4RotationMatrix();
-  placementRotation->rotateX(-CLHEP::halfpi);
   
   BDSExtent ext = CalculateExtents(xBA, xBR, thickness, recipe);
   G4double xLow = ext.XNeg() + xmin;