+++ src/BDSOutputROOTEventExit.cc
@@ -37,16 +37,16 @@ BDSOutputROOTEventExit::~BDSOutputROOTEventExit()
 void BDSOutputROOTEventExit::Fill(const BDSVolumeExitHit* hit)
 {
   n++;
-  totalEnergy.push_back( (float &&) hit->totalEnergy / CLHEP::GeV);
-  postStepKineticEnergy.push_back( (float &&) hit->postStepKineticEnergy / CLHEP::GeV);
-  X.push_back( (float &&) hit->X / CLHEP::m);
-  Y.push_back( (float &&) hit->Y / CLHEP::m);
-  Z.push_back( (float &&) hit->Z / CLHEP::m);
-  T.push_back( (float &&) hit->T / CLHEP::ns);
+  totalEnergy.push_back( (float &&) (hit->totalEnergy / CLHEP::GeV));
+  postStepKineticEnergy.push_back( (float &&) (hit->postStepKineticEnergy / CLHEP::GeV));
+  X.push_back( (float &&) (hit->X / CLHEP::m));
+  Y.push_back( (float &&) (hit->Y / CLHEP::m));
+  Z.push_back( (float &&) (hit->Z / CLHEP::m));
+  T.push_back( (float &&) (hit->T / CLHEP::ns));
   partID.push_back(hit->partID);
   trackID.push_back(hit->trackID);
   parentID.push_back(hit->parentID);
-  weight.push_back( (float &&)  hit->weight);
+  weight.push_back(hit->weight);
   turn.push_back(hit->turnsTaken);
 }
 #endif